Refactored all feature presentations.
