+++ lib/src/features/authentication/presentation/pages/login_page.dart
@@ -1,6 +1,6 @@
 import 'dart:io' show Platform;
 
-import 'package:apk_iman_ba/Pages/registration_page.dart';
+import 'package:apk_iman_ba/src/features/authentication/presentation/pages/registration_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:google_fonts/google_fonts.dart';
@@ -8,7 +8,7 @@ import 'package:the_apple_sign_in/the_apple_sign_in.dart';
 
 import '../../../../../Services/auth_service.dart';
 import '../../../../../components/custom_textfield.dart';
-import '../../../../../pages/forgot_password_page.dart';
+import 'forgot_password_page.dart';
 
 class LoginPage extends StatefulWidget {
   const LoginPage({super.key});
